fix: separate dev menu code form main window code
